blob_trigger_function/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-worker
pandas
azure-storage-blob
azure-ai-textanalytics
pyodbc